Add Skeleton-enemy

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -7,7 +7,7 @@ public abstract class Enemy : MonoBehaviour
     [SerializeField]
     protected int health;
     [SerializeField]
-    protected int speed; 
+    protected float speed; 
     [SerializeField]
     protected int gems;
     [SerializeField]
